Notify stream-creator of stream stop
PipelineImpl posts to the listening response_queue and/or
response_topic when the stream is destroyed.

A stream creator might pass a queue_response or topic
response when calling create_stream(),
so it can be notified as frames are processed.
If a stream exits due to an error in process_frame then these
listeners will be notified, but if the stream exits without
error then the listeners will not previously be notified.